+++ sacred/observers/__init__.py
@@ -7,7 +7,6 @@ from sacred.commandline_options import CommandLineOption
 from sacred.observers.base import RunObserver
 from sacred.observers.file_storage import FileStorageObserver
 import sacred.optional as opt
-from sacred.observers.tinydb_hashfs import TinyDbObserver, TinyDbReader
 
 if opt.has_pymongo:
     from sacred.observers.mongo import MongoObserver
@@ -38,6 +37,25 @@ else:
         def apply(cls, args, run):
             raise ImportError('cannot use -s/--sql flag: '
                               'missing sqlalchemy dependency')
+
+
+if opt.has_tinydb:
+    from sacred.observers.tinydb_hashfs import TinyDbObserver, TinyDbReader
+else:
+    TinyDbObserver = opt.MissingDependencyMock(
+        ['tinydb', 'tinydb-serialization', 'hashfs'])
+    TinyDbReader = opt.MissingDependencyMock(
+        ['tinydb', 'tinydb-serialization', 'hashfs'])
+
+    class TinyDbOption(CommandLineOption):
+        """To use the TinyDBObserver you need to install tinydb first."""
+
+        arg = 'BASEDIR'
+
+        @classmethod
+        def apply(cls, args, run):
+            raise ImportError('cannot use -t/--tiny_db flag: '
+                              'missing tinydb dependency')
 
 
 __all__ = ('FileStorageObserver', 'RunObserver', 'MongoObserver',